setup

+++ plugins/prepare.composer.php
@@ -5,7 +5,13 @@ class PrepareComposer
     {
         $root_directory = realpath(__DIR__.'/..');
         $framework = realpath($root_directory.'/vendor/makscraft/mv-framework');
-        
+
+        if(file_exists($root_directory.'/core/gitkeep'))
+            unlink($root_directory.'/core/gitkeep');
+
+        if(file_exists($root_directory.'/core'))
+            unlink($root_directory.'/core');
+
         $files = [
             '/config/autoload.php',
             '/adminpanel',
@@ -31,7 +37,7 @@ class PrepareComposer
         $dump_old = $root_directory.'/userfiles/database/mysql-dump.sql';
         $dump_new = $root_directory.'/customs/initial-dump.sql';
 
-        if(file_exists($dump_old))
+        if(file_exists($dump_old) && file_exists($dump_new))
             unlink($dump_old);
 
         if(file_exists($dump_new))
